UI test cleanup: Extract or_fun_call tests
